+++ apps/userdashboard/assets/js/a4_candy_userdashboard/ModerationNotificationList.tsx
@@ -3,19 +3,20 @@ import django from 'django'
 
 import { ModerationNotification } from './ModerationNotification'
 import { Filter } from './Filter'
-import type { FilterItem, ModerationComment } from './types'
+import type { FilterItem, ModerationItem } from './types'
 
 const PACKET_COMMENT_SIZE = 15
+
+const contentTypeFilterItems: FilterItem[] = [
+  { label: django.gettext('All'), value: 'all' },
+  { label: django.gettext('All comments'), value: 'comments' },
+  { label: django.gettext('All ideas'), value: 'ideas' }
+]
 
 const isReadFilterItems: FilterItem[] = [
   { label: django.gettext('Read'), value: 'True' },
   { label: django.gettext('Unread'), value: 'False' },
   { label: django.gettext('View all'), value: 'All' }
-]
-
-const reportsFilterItems: FilterItem[] = [
-  { label: django.gettext('Reported'), value: 'True' },
-  { label: django.gettext('All comments'), value: 'All' }
 ]
 
 const orderingFilterItems: FilterItem[] = [
@@ -25,15 +26,20 @@ const orderingFilterItems: FilterItem[] = [
 ]
 
 interface ModerationNotificationListProps {
-  moderationCommentsApiUrl: string
+  moderationCommentsApiUrl?: string
+  moderationItemsApiUrl?: string
   projectTitle: string
   organisation: string
   projectUrl: string
 }
 
 interface ModerationNotificationListState {
-  moderationComments: ModerationComment[]
-  selectedFilters: { isRead: string; hasReports: string; ordering: string }
+  moderationItems: ModerationItem[]
+  selectedFilters: {
+    contentType: string
+    isRead: string
+    ordering: string
+  }
   numOfComments: number
   hasMore: string | null
   packetFactor: number
@@ -49,8 +55,8 @@ export default class ModerationNotificationList extends Component<ModerationNoti
     super(props)
 
     this.state = {
-      moderationComments: [],
-      selectedFilters: { isRead: 'False', hasReports: 'All', ordering: '-num_reports' },
+      moderationItems: [],
+      selectedFilters: { contentType: 'all', isRead: 'False', ordering: '-num_reports' },
       numOfComments: PACKET_COMMENT_SIZE,
       hasMore: null,
       packetFactor: 1,
@@ -63,11 +69,15 @@ export default class ModerationNotificationList extends Component<ModerationNoti
     this.timer = setInterval(() => !this.isLoading && this.loadData(), 3000)
   }
 
-  isReadFilterChangeHandle (value: string) {
+  getItemsApiUrl () {
+    return this.props.moderationItemsApiUrl || this.props.moderationCommentsApiUrl || ''
+  }
+
+  contentTypeFilterChangeHandle (value: string) {
     this.setState({
       selectedFilters: {
         ...this.state.selectedFilters,
-        isRead: value
+        contentType: value
       },
       isLoaded: false
     },
@@ -75,11 +85,11 @@ export default class ModerationNotificationList extends Component<ModerationNoti
     )
   }
 
-  reportsFilterChangeHandle (value: string) {
+  isReadFilterChangeHandle (value: string) {
     this.setState({
       selectedFilters: {
         ...this.state.selectedFilters,
-        hasReports: value
+        isRead: value
       },
       isLoaded: false
     },
@@ -100,8 +110,8 @@ export default class ModerationNotificationList extends Component<ModerationNoti
   }
 
   getUrlParams () {
-    return '?is_reviewed=' + this.state.selectedFilters.isRead +
-      '&has_reports=' + this.state.selectedFilters.hasReports +
+    return '?content_type=' + this.state.selectedFilters.contentType +
+      '&is_reviewed=' + this.state.selectedFilters.isRead +
       '&ordering=' + this.state.selectedFilters.ordering +
       '&num_of_comments=' + this.state.numOfComments
   }
@@ -109,11 +119,11 @@ export default class ModerationNotificationList extends Component<ModerationNoti
   async loadData () {
     this.isLoading = true
     try {
-      const url = this.props.moderationCommentsApiUrl + this.getUrlParams()
+      const url = this.getItemsApiUrl() + this.getUrlParams()
       const data = await fetch(url)
       const jsonData = await data.json()
       this.setState({
-        moderationComments: jsonData.results,
+        moderationItems: jsonData.results,
         hasMore: jsonData.next,
         isLoaded: true
       })
@@ -194,7 +204,7 @@ export default class ModerationNotificationList extends Component<ModerationNoti
     const byText = django.gettext('By ')
     const loadmoreText = django.gettext('Load more')
     const gotoTopText = django.gettext('Go to top')
-    const listText = django.gettext('Notifications of comments from project')
+    const listText = django.gettext('Notifications of comments and ideas from project')
     const filterText = django.gettext('Notification filters and sorting')
     const headerText = django.gettext('Moderation project')
 
@@ -220,9 +230,9 @@ export default class ModerationNotificationList extends Component<ModerationNoti
           <div className="col-md">
             <Filter
               filterClass="filter--full dropdown dropdown-menu-end"
-              filterItems={reportsFilterItems}
-              onFilterChange={(value) => this.reportsFilterChangeHandle(value)}
-              selectedFilter={this.state.selectedFilters.hasReports}
+              filterItems={contentTypeFilterItems}
+              onFilterChange={(value) => this.contentTypeFilterChangeHandle(value)}
+              selectedFilter={this.state.selectedFilters.contentType}
               filterText={django.gettext('Filter')}
             />
           </div>
@@ -256,11 +266,11 @@ export default class ModerationNotificationList extends Component<ModerationNoti
               <div>
                 <h3 id="list-header" className="visually-hidden">{listText}</h3>
                 <ul className="u-list-reset">
-                  {this.state.moderationComments.map((item, i) => (
+                  {this.state.moderationItems.map((item, i) => (
                     <ModerationNotification
                       key={i}
                       notification={item}
-                      apiUrl={this.props.moderationCommentsApiUrl + item.pk + '/'}
+                      apiUrl={item.api_url}
                       getUrlParams={() => this.getUrlParams()}
                       loadData={() => this.loadData()}
                     />