+++ tests/userdashboard/views/test_userdashboard_moderation_detail_view.py
@@ -66,6 +66,11 @@ def test_moderation_dashboard_context_data(client, project_factory):
     )
     assert "moderation_comments_api_url" in context_data
     assert moderation_comments_api_url == context_data["moderation_comments_api_url"]
+    moderation_items_api_url = reverse(
+        "moderationitems-list", kwargs={"project_pk": project.pk}
+    )
+    assert "moderation_items_api_url" in context_data
+    assert moderation_items_api_url == context_data["moderation_items_api_url"]
     assert "back_url" in context_data
     assert context_data["back_url"] == reverse("userdashboard-moderation")
     assert context_data["view"].project == project